CB-14728 Enable stopping clusters with ephemeral based hostgroups based on service role whitelist.
This feature enables stopping clusters with ephemeral volume storage based instance groups. The ephemeral instance group is only permitted to stop if the service roles on it match a given whitelist group. These whitelist groups consist of service roles and roles (service is not specified, only the role) that are not affected by the reformatted ephemeral storage after a stop-start operation. Each service and service role in a group has a required flag, which means that if a service role or role is required, then it has to be present on the instance group to be considered as stoppable. Currently there is onlt one  whitelist group, which contains the YARN NODEMANAGER service role, that is required for the group and a GATEWAY role that is not required. The default DataEngineering templates' "compute" group matches these criterias.