CB-15680. Make cdp-usersync-internal group non-posix
It is significantly faster to add group members to non-posix
groups than to posix groups. The cdp-usersync-internal group
is only used by usersync and does not need to be a posix group.
This commit changes usersync to create cdp-usersync-internal
as non-posix. This will only affect new environments and not
existing cdp-usersync-internal groups.